+++ internal/kademlia/model/model.go
@@ -2,7 +2,7 @@ package model
 
 import "d7024e_group04/internal/kademlia/contact"
 
-type ValueObject struct {
+type FindValueSuccessfulResponse struct {
 	DataValue     string
 	NodeWithValue *contact.Contact
 }
